Zhao Figure 4
As in Zhao and my previous analysis, perturbation width of 100 K, and magnitude of 0.05 is supported. At very low width, the sensitivity behaves erratically, I believe this is caused by grid spacing. For narrow perturbations, there may not be a grid point close enough to the perturbation temperature to be affected.